+++ src/commands/plugins/install.ts
@@ -1,8 +1,8 @@
 import {Command, flags} from '@oclif/command'
-import * as chalk from 'chalk'
+import chalk = require('chalk')
 import cli from 'cli-ux'
 
-import Plugins from '../../plugins'
+import Plugins from '../../modules/plugins'
 
 export default class PluginsInstall extends Command {
   static description = `installs a plugin into the CLI